requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
nats-py